requirements.txt
APScheduler
asgiref
asttokens
azure-common
azure-core
azure-storage-blob
bcrypt
blinker
boto3
botocore
cachetools
certifi
cffi
charset-normalizer
click
colorama
comm
cryptography
debugpy
decorator
defusedxml
distlib
Django
executing
filelock
Flask
Flask-Cors
google-ads
google-ads-api-report-fetcher
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
gunicorn
httpie
httplib2
idna
ipykernel
ipython
isodate
itsdangerous
jedi
Jinja2
jmespath
jupyter_client
jupyter_core
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
multidict
nest-asyncio
numpy
oauth2client
oauthlib
packaging
pandas
paramiko
parso
pipenv
platformdirs
prompt-toolkit
proto-plus
protobuf
psutil
pure-eval
pyarrow
pyasn1
pyasn1-modules
pycparser
Pygments
PyJWT
PyNaCl
pyparsing
PySocks
python-dateutil
pytz
PyYAML
pyzmq
requests
requests-oauthlib
requests-toolbelt
rich
rsa
s3transfer
secure-smtplib
six
smart-open
sqlparse
stack-data
tenacity
tornado
traitlets
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
virtualenv
waitress
wcwidth
Werkzeug